5.0 - alpha
完善词典设置、词典笔记。

use long instead of int.

完善全文搜索

+++ PLOD/src/main/java/com/knziha/plod/dictionary/GetRecordAtInterceptor.java
@@ -1,5 +1,5 @@
 package com.knziha.plod.dictionary;
 
 public interface GetRecordAtInterceptor {
-	String getRecordAt(UniversalDictionaryInterface bookImpl, int position);
+	String getRecordAt(UniversalDictionaryInterface bookImpl, long position);
 }
